Refactor SendSearchResult to use Factory Methods

+++ src/CleanReaderBot.Application/Common/Entities/Book.cs
@@ -5,7 +5,7 @@ namespace CleanReaderBot.Application.Common.Entities
         public int Id { get; set; }
         public string Title { get; set; }
         public Author Author { get; set; }
-        public float AverageRating { get; set; }
+        public double AverageRating { get; set; }
         public string ImageUrl { get; set; }
         public string SmallImageUrl { get; set; }
     }
